move scheduler
